+++ src/realm-settings/AESGeneratedModal.tsx
@@ -22,6 +22,7 @@ import type ComponentRepresentation from "keycloak-admin/lib/defs/componentRepre
 import { HelpItem } from "../components/help-enabler/HelpItem";
 import { useServerInfo } from "../context/server-info/ServerInfoProvider";
 import { useRealm } from "../context/realm-context/RealmContext";
+import { AESGeneratedForm } from "./AESGeneratedForm";
 
 type AESGeneratedModalProps = {
   providerType?: string;
@@ -100,153 +101,7 @@ AESGeneratedModalProps) => {
         </Button>,
       ]}
     >
-      <Form
-        isHorizontal
-        id="add-provider"
-        className="pf-u-mt-lg"
-        onSubmit={handleSubmit(save!)}
-      >
-        <FormGroup
-          label={t("realm-settings:consoleDisplayName")}
-          fieldId="kc-console-display-name"
-          labelIcon={
-            <HelpItem
-              helpText="realm-settings-help:displayName"
-              forLabel={t("loginTheme")}
-              forID="kc-console-display-name"
-            />
-          }
-        >
-          <Controller
-            name="name"
-            control={control}
-            defaultValue={providerType}
-            render={({ onChange }) => (
-              <TextInput
-                aria-label={t("consoleDisplayName")}
-                defaultValue={providerType}
-                onChange={(value) => {
-                  onChange(value);
-                  setDisplayName(value);
-                }}
-                data-testid="display-name-input"
-              ></TextInput>
-            )}
-          />
-        </FormGroup>
-        <FormGroup
-          label={t("common:enabled")}
-          fieldId="kc-enabled"
-          labelIcon={
-            <HelpItem
-              helpText={t("realm-settings-help:enabled")}
-              forLabel={t("enabled")}
-              forID="kc-enabled"
-            />
-          }
-        >
-          <Controller
-            name="config.enabled"
-            control={control}
-            defaultValue={["true"]}
-            render={({ onChange, value }) => (
-              <Switch
-                id="kc-enabled"
-                label={t("common:on")}
-                labelOff={t("common:off")}
-                isChecked={value[0] === "true"}
-                data-testid={
-                  value[0] === "true"
-                    ? "internationalization-enabled"
-                    : "internationalization-disabled"
-                }
-                onChange={(value) => {
-                  onChange([value + ""]);
-                }}
-              />
-            )}
-          />
-        </FormGroup>
-        <FormGroup
-          label={t("realm-settings:active")}
-          fieldId="kc-active"
-          labelIcon={
-            <HelpItem
-              helpText="realm-settings-help:active"
-              forLabel={t("active")}
-              forID="kc-active"
-            />
-          }
-        >
-          <Controller
-            name="config.active"
-            control={control}
-            defaultValue={["true"]}
-            render={({ onChange, value }) => (
-              <Switch
-                id="kc-active"
-                label={t("common:on")}
-                labelOff={t("common:off")}
-                isChecked={value[0] === "true"}
-                data-testid={
-                  value[0] === "true"
-                    ? "internationalization-enabled"
-                    : "internationalization-disabled"
-                }
-                onChange={(value) => {
-                  onChange([value + ""]);
-                }}
-              />
-            )}
-          />
-        </FormGroup>
-        {providerType === "aes-generated" && (
-          <FormGroup
-            label={t("realm-settings:AESKeySize")}
-            fieldId="kc-aes-keysize"
-            labelIcon={
-              <HelpItem
-                helpText="realm-settings-help:AESKeySize"
-                forLabel={t("AESKeySize")}
-                forID="kc-aes-key-size"
-              />
-            }
-          >
-            <Controller
-              name="config.secretSize"
-              control={control}
-              defaultValue={["16"]}
-              render={({ onChange, value }) => (
-                <Select
-                  toggleId="kc-aes-keysize"
-                  onToggle={() =>
-                    setIsKeySizeDropdownOpen(!isKeySizeDropdownOpen)
-                  }
-                  onSelect={(_, value) => {
-                    onChange([value + ""]);
-                    setIsKeySizeDropdownOpen(false);
-                  }}
-                  selections={[value + ""]}
-                  isOpen={isKeySizeDropdownOpen}
-                  variant={SelectVariant.single}
-                  aria-label={t("aesKeySize")}
-                  data-testid="select-secret-size"
-                >
-                  {allComponentTypes[0].properties[3].options!.map(
-                    (item, idx) => (
-                      <SelectOption
-                        selected={item === value}
-                        key={`email-theme-${idx}`}
-                        value={item}
-                      />
-                    )
-                  )}
-                </Select>
-              )}
-            />
-          </FormGroup>
-        )}
-      </Form>
+      <AESGeneratedForm  save={save} providerType={providerType}/>
     </Modal>
   );
 };
