+++ src/realm-settings/RSAGeneratedForm.tsx
@@ -1,0 +1,239 @@
+import React, { useState } from "react";
+import {
+  AlertVariant,
+  Button,
+  ButtonVariant,
+  Form,
+  FormGroup,
+  Modal,
+  ModalVariant,
+  Select,
+  SelectOption,
+  SelectVariant,
+  Switch,
+  TextInput,
+} from "@patternfly/react-core";
+import { useTranslation } from "react-i18next";
+import { Controller, useForm } from "react-hook-form";
+
+import { useAdminClient } from "../context/auth/AdminClient";
+import { useAlerts } from "../components/alert/Alerts";
+import type ComponentRepresentation from "keycloak-admin/lib/defs/componentRepresentation";
+import { HelpItem } from "../components/help-enabler/HelpItem";
+import { useServerInfo } from "../context/server-info/ServerInfoProvider";
+import { useRealm } from "../context/realm-context/RealmContext";
+
+type RSAGeneratedFormProps = {
+  providerType?: string;
+  handleModalToggle?: () => void;
+  refresh?: () => void;
+  save: (component: ComponentRepresentation) => void;
+};
+
+export const RSAGeneratedForm = ({
+  providerType,
+  handleModalToggle,
+  save,
+}: RSAGeneratedFormProps) => {
+  const { t } = useTranslation("groups");
+  const serverInfo = useServerInfo();
+  const { handleSubmit, control } = useForm({});
+  const [isKeySizeDropdownOpen, setIsKeySizeDropdownOpen] = useState(false);
+  const [isRSAalgDropdownOpen, setIsRSAalgDropdownOpen] = useState(false);
+  const [displayName, setDisplayName] = useState("");
+
+  const allComponentTypes = serverInfo.componentTypes![
+    "org.keycloak.keys.KeyProvider"
+  ];
+
+  return (
+    <Form
+      isHorizontal
+      id="add-provider"
+      className="pf-u-mt-lg"
+      onSubmit={handleSubmit(save!)}
+    >
+      <FormGroup
+        label={t("realm-settings:consoleDisplayName")}
+        fieldId="kc-console-display-name"
+        labelIcon={
+          <HelpItem
+            helpText="realm-settings-help:displayName"
+            forLabel={t("loginTheme")}
+            forID="kc-console-display-name"
+          />
+        }
+      >
+        <Controller
+          name="name"
+          control={control}
+          defaultValue={providerType}
+          render={({ onChange }) => (
+            <TextInput
+              aria-label={t("consoleDisplayName")}
+              defaultValue={providerType}
+              onChange={(value) => {
+                onChange(value);
+                setDisplayName(value);
+              }}
+              data-testid="display-name-input"
+            ></TextInput>
+          )}
+        />
+      </FormGroup>
+      <FormGroup
+        label={t("common:enabled")}
+        fieldId="kc-enabled"
+        labelIcon={
+          <HelpItem
+            helpText={t("realm-settings-help:enabled")}
+            forLabel={t("enabled")}
+            forID="kc-enabled"
+          />
+        }
+      >
+        <Controller
+          name="config.enabled"
+          control={control}
+          defaultValue={["true"]}
+          render={({ onChange, value }) => (
+            <Switch
+              id="kc-enabled"
+              label={t("common:on")}
+              labelOff={t("common:off")}
+              isChecked={value[0] === "true"}
+              data-testid={
+                value[0] === "true"
+                  ? "internationalization-enabled"
+                  : "internationalization-disabled"
+              }
+              onChange={(value) => {
+                onChange([value + ""]);
+              }}
+            />
+          )}
+        />
+      </FormGroup>
+      <FormGroup
+        label={t("realm-settings:active")}
+        fieldId="kc-active"
+        labelIcon={
+          <HelpItem
+            helpText="realm-settings-help:active"
+            forLabel={t("active")}
+            forID="kc-active"
+          />
+        }
+      >
+        <Controller
+          name="config.active"
+          control={control}
+          defaultValue={["true"]}
+          render={({ onChange, value }) => (
+            <Switch
+              id="kc-active"
+              label={t("common:on")}
+              labelOff={t("common:off")}
+              isChecked={value[0] === "true"}
+              data-testid={
+                value[0] === "true"
+                  ? "internationalization-enabled"
+                  : "internationalization-disabled"
+              }
+              onChange={(value) => {
+                onChange([value + ""]);
+              }}
+            />
+          )}
+        />
+      </FormGroup>
+        <>
+          <FormGroup
+            label={t("realm-settings:algorithm")}
+            fieldId="kc-algorithm"
+            labelIcon={
+              <HelpItem
+                helpText="realm-settings-help:algorithm"
+                forLabel={t("algorithm")}
+                forID="kc-algorithm"
+              />
+            }
+          >
+            <Controller
+              name="algorithm"
+              defaultValue=""
+              render={({ onChange, value }) => (
+                <Select
+                  toggleId="kc-rsa-algorithm"
+                  onToggle={() =>
+                    setIsRSAalgDropdownOpen(!isRSAalgDropdownOpen)
+                  }
+                  onSelect={(_, value) => {
+                    onChange(value as string);
+                    setIsRSAalgDropdownOpen(false);
+                  }}
+                  selections={[value + ""]}
+                  variant={SelectVariant.single}
+                  aria-label={t("algorithm")}
+                  isOpen={isRSAalgDropdownOpen}
+                  data-testid="select-rsa-algorithm"
+                >
+                  {allComponentTypes[5].properties[3].options!.map((p, idx) => (
+                    <SelectOption
+                      selected={p === value}
+                      key={`rsa-algorithm-${idx}`}
+                      value={p}
+                    ></SelectOption>
+                  ))}
+                </Select>
+              )}
+            />
+          </FormGroup>
+          <FormGroup
+            label={t("realm-settings:AESKeySize")}
+            fieldId="kc-aes-keysize"
+            labelIcon={
+              <HelpItem
+                helpText="realm-settings-help:AESKeySize"
+                forLabel={t("AESKeySize")}
+                forID="kc-aes-key-size"
+              />
+            }
+          >
+            <Controller
+              name="config.secretSize"
+              control={control}
+              defaultValue={["2048"]}
+              render={({ onChange, value }) => (
+                <Select
+                  toggleId="kc-rsa-keysize"
+                  onToggle={() =>
+                    setIsKeySizeDropdownOpen(!isKeySizeDropdownOpen)
+                  }
+                  onSelect={(_, value) => {
+                    onChange([value + ""]);
+                    setIsKeySizeDropdownOpen(false);
+                  }}
+                  selections={[value + ""]}
+                  isOpen={isKeySizeDropdownOpen}
+                  variant={SelectVariant.single}
+                  aria-label={t("keySize")}
+                  data-testid="select-secret-size"
+                >
+                  {allComponentTypes[5].properties[4].options!.map(
+                    (item, idx) => (
+                      <SelectOption
+                        selected={item === value}
+                        key={`rsa-generated-key-size-${idx}`}
+                        value={item}
+                      />
+                    )
+                  )}
+                </Select>
+              )}
+            />
+          </FormGroup>
+        </>
+    </Form>
+  );
+};
